requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.4.2 
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
alembic==1.7.1
motor==3.3.1
pymongo==4.6.1
redis==5.0.1 